+++ src/components/sidebar.tsx
@@ -220,7 +220,7 @@ function Sidebar({
 
     return (
         <div
-            className="group peer text-text-primary hidden md:block"
+            className="group peer text-text-primary hidden lg:block"
             data-state={state}
             data-collapsible={state === 'collapsed' ? collapsible : ''}
             data-variant={variant}